📝 Add .envrc file to set OCO_LANGUAGE environment variable 📝 Update README.md with a screenshot of the application 🔧 Add assets/image.png file containing the screenshot
The changes introduce a new .envrc file to set the OCO_LANGUAGE environment variable, likely for internationalization purposes. The README.md file has been updated with a new screenshot of the application, and a corresponding assets/image.png file has been added to the repository. These changes improve the documentation and provide a visual representation of the application.