🔧 (helm/templates/binding.yaml): update ClusterRole rules to allow get, list and watch on all resources in kustomize.toolkit.fluxcd.io API group
🐛 (helm/values.yaml): fix indentation for application.ports and ingress.hosts to conform to YAML syntax

✨ (helm/values.yaml): add support for deploying multiple ingress hosts by changing ingress.hosts from a string to a list

🐛 (helm/values.yaml): fix typo in serviceAccount.name from fluxtcd-viewer to fluxcd-viewer

✨ (taskfile.yaml): add new task 'deploy' to deploy helm chart using 'helm upgrade --install'

The changes in helm/templates/binding.yaml update the ClusterRole rules to allow more permissions for the fluxcd-viewer application. The indentation fixes in helm/values.yaml ensure the YAML is valid. Support for multiple ingress hosts is added by changing ingress.hosts to a list. A typo in the serviceAccount name is also fixed. Finally, a new 'deploy' task is added to taskfile.yaml to simplify deploying the helm chart.